app/requirements.txt
numpy==1.26.2
streamlit==1.29.0
streamlit-option-menu==0.3.6
tensorflow==2.15.0.post1
Pillow==10.1.0
